git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@4640 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/ANALYSIS/MAPMATCHING/PoseClusteringAffineSuperimposer.h
@@ -4,7 +4,7 @@
 // --------------------------------------------------------------------------
 //                   OpenMS Mass Spectrometry Framework
 // --------------------------------------------------------------------------
-//  Copyright (C) 2003-2008 -- Oliver Kohlbacher, Knut Reinert
+//  Copyright (C) 2003-2009 -- Oliver Kohlbacher, Knut Reinert
 //
 //  This library is free software; you can redistribute it and/or
 //  modify it under the terms of the GNU Lesser General Public
@@ -183,7 +183,7 @@ namespace OpenMS
 						++i
 					)
 			{
-				setProgress(10+Real(i)/model_map_size*100.f);
+				setProgress((UInt)(10+Real(i)/model_map_size*100.f));
 
 				// Adjust window in scene map
 				while ( k_low < scene_map_size &&
